Rename `db.models` to `models`
