+++ multilingual/__init__.py
@@ -10,6 +10,6 @@ __version__ = '0.3.1'
 
 
 from .admin import MultilingualModelAdmin
-from .db.models.base import MultilingualModel
-from .db.models.manager import MultilingualManager
 from .forms import MultilingualModelForm
+from .models.base import MultilingualModel
+from .models.manager import MultilingualManager
