--- multilingual/db/__init__.py
@@ -1,3 +1,0 @@
-"""
-Alters django db core to enable existence of multilingual models
-"""
