--- multilingual/db/models/__init__.py
@@ -1,3 +1,0 @@
-"""
-Support for multilingual models
-"""
