--- multilingual/models.py
@@ -1,3 +1,0 @@
-"""
-File required by django to enable tests.
-"""
