+++ multilingual/models/__init__.py
@@ -1,0 +1,5 @@
+"""
+Base classes for multilingual models.
+
+Structure of this directory corresponds to `django.db.models` directory.
+"""
