+++ multilingual/models/query.py
@@ -3,7 +3,7 @@ Queryset for multilingual models
 """
 from django.db.models.query import QuerySet
 
-from multilingual.db.models.sql.query import MultilingualQuery
+from .sql.query import MultilingualQuery
 
 
 class MultilingualQuerySet(QuerySet):
